requirements.txt
streamlit
matplotlib
folium
pandas
seaborn
numpy
plotly
pyportfolioopt
Pillow
statsmodels
sklearn


